chore: CardList 컴포넌트 스타일코드 분리
